requirements.txt
flask==3.0.0
python-dotenv==1.0.0
mistralai==0.0.7
SQLAlchemy==2.0.23
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0
typing_extensions==4.8.0
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
urllib3==2.1.0
